+++ menu/drivers/xmb.c
@@ -3894,7 +3894,7 @@ static void xmb_render(void *data,
 
             if ((pointer_y > y1) && (pointer_y < y2))
             {
-               menu_input_set_pointer_selection(i);
+               menu_input_set_pointer_selection((unsigned)i);
                break;
             }
          }
